Update bar chart values loading in ordered way

+++ model/metric.go
@@ -3,9 +3,9 @@ package model
 // SystemMetric basic metric data of observable system
 type SystemMetric struct {
 	// RequestCount of system
-	RequestCount int
+	RequestCount float32
 	// ErrorCount of system
-	ErrorCount int
+	ErrorCount float32
 }
 
 // SystemMetaData additional information in string format
